Crate bl602_sdk

Rust Wrapper for BL602 IoT SDK. See “The RISC-V BL602 Book” https://lupyuen.github.io/articles/book

Modules

adc

ADC HAL for BL602. See https://lupyuen.github.io/articles/book#adc-on-bl602

dma

DMA HAL for BL602. See https://lupyuen.github.io/articles/book#dma-on-bl602

gpio

GPIO HAL for BL602. See https://lupyuen.github.io/articles/book#gpio-on-bl602

i2c

I2C HAL for BL602. See https://lupyuen.github.io/articles/book#i2c-on-bl602

pwm

PWM HAL for BL602. See https://lupyuen.github.io/articles/book#pwm-on-bl602

result

HAL return type and error codes

spi

SPI HAL for BL602. See https://lupyuen.github.io/articles/book#spi-on-bl602

uart

UART HAL for BL602. See https://lupyuen.github.io/articles/book#uart-on-bl602

wifi

WiFi HAL for BL602. See https://lupyuen.github.io/articles/book#wifi-on-bl602

Structs

Strn

Represents a null-terminated string, suitable for passing to C APIs as * const char. The string can be a null-terminated byte string created in Rust, or a pointer to a null-terminated string returned by C. Pointer may be null.

Enums

StrnRep

Either a byte string or a string pointer

Functions

puts

Print a message to the serial console. TODO: Auto-generate this wrapper with bindgen from the C declaration

time_delay

Sleep for the specified number of system ticks. TODO: Auto-generate this wrapper with bindgen from the C declaration: void ble_npl_time_delay(ble_npl_time_t ticks)

time_ms_to_ticks32

Convert milliseconds to system ticks. TODO: Auto-generate this wrapper with bindgen from the C declaration: ble_npl_time_t ble_npl_time_ms_to_ticks32(uint32_t ms)

Type Definitions

Ptr

Declare a void * pointer that will be passed to C functions

String

Limit Strings to 64 chars, similar to char[64] in C